Some comments

+++ src/main.c
@@ -18,11 +18,12 @@ int main(void) {
 
     const int texture_x = SCREEN_WIDTH / 2 - texture.width / 2;
     const int texture_y = SCREEN_HEIGHT / 2 - texture.height / 2;
+    // draw the background of the window
     DrawTexture(texture, texture_x, texture_y, WHITE);
 
     const char *text = "BOTTOM TEXT";
     const Vector2 text_size = MeasureTextEx(GetFontDefault(), text, 20, 1);
-    DrawText(text, SCREEN_WIDTH / 2 - text_size.x / 2,
+    DrawText(text, (int) SCREEN_WIDTH / 2 -  (int) text_size.x / 2,
              texture_y + texture.height + text_size.y + 10, 20, BLACK);
 
     EndDrawing();
